feat: make url list a client side component
